# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: MasterEat API
  version: 1.0.0
  description: |
    Specifica OpenAPI delle API REST implementate in MasterEat.

    Le API sotto `/api/*` sono protette da un filtro CORS dedicato che abilita
    anche client JavaScript esterni. Le richieste preflight `OPTIONS` ricevono
    `204 No Content`; gli header ammessi sono `Authorization`, `Content-Type`
    e `Accept`. Non è previsto supporto a credenziali cross-origin basate su
    cookie.

    Il login restituisce un token interno. Gli endpoint protetti richiedono
    l'header `Authorization: Bearer <token>`. Il codice accetta anche l'header
    `Authentication` con il token non prefissato, ma `Authorization` è il
    formato documentato.

    Gli stati ordine gestiti sono `BOZZA`, `INSERITO`, `IN_PREPARAZIONE`,
    `PRONTO`, `IN_CONSEGNA`, `CONSEGNATO`, `ANNULLATO`. Le transizioni
    operative sono progressive: `INSERITO -> IN_PREPARAZIONE -> PRONTO ->
    IN_CONSEGNA -> CONSEGNATO`.

    `DELETE /ordini/{idOrdine}` effettua un annullamento logico impostando lo
    stato ad `ANNULLATO`, non una cancellazione fisica.
servers:
  - url: http://localhost:8080/MasterEat/api
    description: Server locale Tomcat con WAR MasterEat e servlet mapping /api/*
...
paths:
...
  /prodotti/{idProdotto}/ingredienti:
    get:
      tags:
        - Ingredienti
      summary: 10 - Ingredienti di un prodotto
      description: |
        Endpoint protetto: richiede token `Authorization: Bearer <token>`.
        Ruoli ammessi: `PERSONALE`, `PROPRIETARIO`. Restituisce gli ingredienti
        attivi associati al prodotto e la quantità usata.
      operationId: getIngredientiByProdottoId
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdProdottoPath'
      responses:
        '200':
          description: Ingredienti del prodotto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientiProdottoResponse'
              examples:
                margherita:
                  value:
                    idProdotto: 1
                    nomeProdotto: Pizza Margherita
                    ingredienti:
                      - id: 1
                        nome: Farina
                        quantita: 250.000
                        unitaMisura: g
                      - id: 2
                        nome: Pomodoro
                        quantita: 80.000
                        unitaMisura: g
                      - id: 3
                        nome: Mozzarella
                        quantita: 100.000
                        unitaMisura: g
                      - id: 4
                        nome: Basilico
                        quantita: 5.000
                        unitaMisura: g
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internal-token
      description: "Token interno restituito da POST /auth/login, inviato come `Authorization: Bearer <token>`."
  parameters:
...
    IdProdottoPath:
      name: idProdotto
      in: path
      required: true
      description: Identificativo del prodotto.
      schema:
        type: integer
        format: int32
        minimum: 1
      example: 1
  responses:
    BadRequest:
      description: Richiesta non valida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            errore:
              value:
                errore: Parametro non valido
    Unauthorized:
      description: Token mancante, scaduto o non valido, oppure credenziali non valide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            errore:
              value:
                errore: Token mancante o non valido
    Forbidden:
      description: Utente autenticato ma non autorizzato per ruolo o proprietà della risorsa.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            errore:
              value:
                errore: Operazione consentita solo a PERSONALE o PROPRIETARIO
    NotFound:
      description: Risorsa non trovata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            errore:
              value:
                errore: Risorsa non trovata
    InternalServerError:
      description: Errore interno.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            errore:
              value:
                errore: Errore interno
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - errore
      properties:
        errore:
          type: string
          example: Token mancante o non valido
...
    IngredientiProdottoResponse:
      type: object
      required:
        - idProdotto
        - nomeProdotto
        - ingredienti
      properties:
        idProdotto:
          type: integer
          format: int64
        nomeProdotto:
          type: string
        ingredienti:
          type: array
          items:
            $ref: '#/components/schemas/IngredienteProdottoResponse'
    IngredienteProdottoResponse:
      type: object
      required:
        - id
        - nome
        - quantita
        - unitaMisura
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        quantita:
          type: number
          format: double
          minimum: 0
        unitaMisura:
          type: string
...